refactor: organise config folder
