+++ backend/open_webui/config/__init__.py
@@ -1,20 +1,5 @@
-from .base import *
-from .logging import *
-from .database import *
-
-from .llm import *
-from .auth import *
-from .media import *
-from .retrieval import *
-from .vector import *
-from .code_interpreter import *
-from .storage import *
-from .admin import *
-from .ui import *
-from .tasks import *
-from .static import *
-from .permissions import *
-from .misc import *
+from .core import *
+from .integrations import *
 from .features import *
 
 setup_logging()
