+++ backend/open_webui/config/core/__init__.py
@@ -1,0 +1,8 @@
+"""
+Core system configurations
+"""
+
+from .base import *
+from .database import *
+from .auth import *
+from .logging import *
