+++ backend/open_webui/config/features/__init__.py
@@ -1,0 +1,11 @@
+"""
+Feature specific configurations
+"""
+
+from .ui import *
+from .admin import *
+from .permissions import *
+from .tasks import *
+from .static import *
+from .misc import *
+from .features import *
