+++ backend/open_webui/config/features/static.py
