+++ backend/open_webui/config/features/tasks.py
@@ -1,5 +1,5 @@
 import os
-from open_webui.config.base import PersistentConfig
+from open_webui.config.core.base import PersistentConfig
 
 ####################################
 # TASK CONFIGURATION
