+++ backend/open_webui/config/integrations/__init__.py
@@ -1,0 +1,10 @@
+"""
+External service integrations
+"""
+
+from .llm import *
+from .vector import *
+from .storage import *
+from .media import *
+from .retrieval import *
+from .code_interpreter import *
