+++ backend/open_webui/config/integrations/code_interpreter.py
@@ -1,5 +1,5 @@
 import os
-from open_webui.config.base import PersistentConfig
+from open_webui.config.core.base import PersistentConfig
 
 
 ENABLE_CODE_EXECUTION = PersistentConfig(
