+++ backend/open_webui/config/integrations/vector.py
